tidy up the mapping classes, adding a base
