+++ app/models/group.rb
@@ -88,6 +88,14 @@ class Group < ApplicationRecord
     where(category: true).where.not(name: MAJOR_POLITICAL_CATEGORIES).order(:name)
   end
 
+  scope :political_parties, -> do
+    where(
+      id: Membership.joins(:group)
+                    .where(groups: { category: true, name: MAJOR_POLITICAL_CATEGORIES })
+                    .select(:member_id)
+    )
+  end
+
   def business_number=(value)
     return if value.nil?
